restructure fragments
