+++ src/app/src/main/java/com/example/myapplication/ui/scanQR/ScanQRViewModel.java
@@ -1,7 +1,6 @@
 package com.example.myapplication.ui.scanQR;
 
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
+
 import androidx.lifecycle.ViewModel;
 
 import com.example.myapplication.Event;
