+++ src/app/src/main/res/layout/fragment_search_event.xml
@@ -5,8 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/appBackground"
-    tools:context=".ui.scanQR.SearchEventFragment"
-    >
+    tools:context=".ui.scanQR.SearchEventFragment">
 
     <TextView
         android:id="@+id/notify_message"
