+++ src/app/src/main/res/layout/scan_qr_screen_fragment.xml
@@ -1,69 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
     android:background="@color/appBackground"
-    tools:context=".ui.scanQR.ScanQRFragment">
-
-    <TextView
-        android:id="@+id/scan_qr_header"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="50dp"
-        android:text="Scan QR Code"
-        android:textAlignment="center"
-        android:textColor="@color/black"
-        android:textSize="30sp"
-        android:textStyle="bold"
-        tools:ignore="HardcodedText" />
-
-    <ImageView
-        android:id="@+id/qr_image_demo"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:src="@drawable/qrdemo"
-        tools:ignore="ContentDescription" />
-
-    <LinearLayout
-        android:id="@+id/info_frame"
-        android:layout_width="match_parent"
-        android:layout_marginStart="50dp"
-        android:layout_height="60dp"
-        android:layout_marginTop="40dp"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="50dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="0"
-            android:src="@android:drawable/ic_menu_info_details"
-            tools:ignore="ContentDescription,InefficientWeight" />
-
-        <TextView
-            android:id="@+id/scan_instructions"
-            android:layout_width="248dp"
-            android:layout_height="wrap_content"
-            android:text="Click the button to find event details and to enroll"
-            android:textAlignment="viewStart"
-            android:textStyle="bold"
-            tools:ignore="HardcodedText,TextSizeCheck" />
-    </LinearLayout>
-
-    <Button
-        android:id="@+id/scan_button"
-        android:layout_width="252dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginTop="0dp"
-        android:layout_marginBottom="100dp"
-        android:backgroundTint="@color/defaultButtonColor"
-        android:text="Scan QR Code"
-        app:cornerRadius="6dp"
-        tools:ignore="HardcodedText,DuplicateSpeakableTextCheck" />
-</LinearLayout>
+    tools:context=".ui.scanQR.ScanQRFragment" />
